+++ src/DocumentEngineBase.sol
@@ -1,5 +1,5 @@
 //SPDX-License-Identifier: MPL-2.0
-pragma solidity ^0.8.20;
+pragma solidity ^0.8.24;
 
 import "OZ/utils/Context.sol";
 import "CMTAT/interfaces/tokenization/draft-IERC1643.sol";